requirements.txt
boto3==1.35.27
python-dotenv==1.0.1
supabase==2.7.4